Make config object more 'dynamic'
Before now the values of the config object are read from
/sys/server/config.json
Now it is possible for some values (like 'port' and 'host')
to be set dynamically.
This is useful before certain environments like Heroku
will not run the Bolt server on predetermined ports,
but would rather supply their own ports.