Remove 'ini' and 'install'
The fields bolt.ini and bolt.install have been removed.
This is because Bolt now raises events like
"bolt/app-starting" and "bolt/app-installed"
which apps can listen for to know when they've been started or
installed.

What's more. The event "bolt/app-starting" (and "bolt/app-stopping")
are targeted events: they only get sent to the app that is about to
start or stop.