Add little delay
I added a little delay between when 'app-starting' event is fired
and when 'app-started' event is fired.
The delay is meant to give the app ample time to start.
This is not the ideal way to do this but our event-dispatching model,
which doesn't tell us if each subscriber has actually handled an event,
leaves us with little choice other than this.

This same has not (yet) been done for 'app-stopping' and 'app-stopped'
because that issue (giving app time to stop) isn't a top priority issue
with Bolt currently.