No more support for 'bolt.startup' in package.json
Bolt no longer supports the startup field, which determined if an
app was started as soon as Bolt started.
The reason for this is that, irrespective of what the end user chooses,
developers can easily achieve this behaviour by listening for events
that are fired while Bolt is booting, like 'bolt/system-started'.