Check for invalid characters in app name
From now on you won't be able to install apps whose names have
characters like '/', '\', '?', '&', ...
That is, names that are not URL-friendly.
This also helps us during the name resolution of app collections.
The full name of collections created by apps will now be:
{app name}/{collection name}
(as against {app name}-{collection name})
This way we are sure 2 collections can never have name conflicts.

In the former method you could have name conflicts if
app 'a-b' has a collection named 'c' and
app 'a' has a collection named 'b-c' because both collections will
have the full name 'a-b-c'.
Now that is no longer possible since the character used in
concatenating app name and collection name ('/') cannot appear in
an app name (you can't have an app named 'a/b').
(The character can't also appear in collection name, but we don't
check for that; we allow you create such offending collections but
you may not be able to access them in a URL like '/api/db/...').